requirements.txt
boto3==1.9.23
mysql-replication==0.18
redis==2.10.6
pika==0.12.0
